webview js: extract `appendAuthToImages` to module
Extract `appendAuthToImages` to a module, in preparation for writing
tests for it.

No functional changes.
