+++ src/webview/js/js.js
@@ -11,6 +11,8 @@ import type {
 } from '../webViewHandleUpdates';
 import type { MessageListEvent } from '../webViewEventHandlers';
 
+import appendAuthToImages from './appendAuthToImages';
+
 /*
  * Supported platforms:
  *
@@ -477,32 +479,6 @@ const scrollToPreserve = (msgId: number, prevBoundTop: number) => {
   }
   const newBoundRect = newElement.getBoundingClientRect();
   window.scrollBy(0, newBoundRect.top - prevBoundTop);
-};
-
-const appendAuthToImages = auth => {
-  const imageTags = document.getElementsByTagName('img');
-  Array.from(imageTags).forEach(img => {
-    if (!img.src.startsWith(auth.realm)) {
-      return;
-    }
-
-    // This unusual form of authentication is only accepted by the server
-    // for a small number of routes.  Rather than append the API key to all
-    // kinds of URLs on the server, do so only for those routes.
-    const srcPath = img.src.substring(auth.realm.length);
-    if (
-      !(
-        srcPath.startsWith('/user_uploads/')
-        || srcPath.startsWith('/thumbnail?')
-        || srcPath.startsWith('/avatar/')
-      )
-    ) {
-      return;
-    }
-
-    const delimiter = img.src.includes('?') ? '&' : '?';
-    img.src += `${delimiter}api_key=${auth.apiKey}`;
-  });
 };
 
 const handleUpdateEventContent = (uevent: WebViewUpdateEventContent) => {
